Solution to Simple Calculator

+++ Simple Calculator@@6788dd63490637cfc44e1ba8/code.c
@@ -6,12 +6,12 @@ int main() {
     int sum = n1 + n2;
     int sub = n1 - n2;
     int multiply = n1 * n2;
-    int div = n1 / n2;
+    float div = n1 / n2;
     switch (ch){
          case '+':printf("%d\n",sum);break;
          case '-':printf("%d\n",sub);break;
          case '*':printf("%d\n",multiply);break;
-         case '/':printf("%d\n",div);break;
+         case '/':printf("%f\n",div);break;
          default:printf("error");}
          return 0;
 }
